+++ srcjs/src/types/index.ts
@@ -217,6 +217,7 @@ export interface InteractionSpec {
   enableSelect: boolean;
   enableHover: boolean;
   enableResize: boolean;
+  enableExport?: boolean;
 }
 
 export interface LayoutSpec {